% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euro_nuts2.R
\name{euro_nuts2}
\alias{euro_nuts2}
\title{Loads a NUTS2-level map of Europe, for plotting Eurostat data}
\usage{
euro_nuts2(keep = 0.05, region = "NUTS_ID")
}
\arguments{
\item{keep}{The amount you want to simplify the shapefile to - defaults to five percent}

\item{region}{What you want to call each region when fortified. This should be an area code or other unique identifier.
It defaults to the 4-letter NUTS ID}
}
\description{
This function loads a map of European NUTS2 regions, ready for plotting with ggplot.
}
\examples{
europe <- euro_nuts2() \%>\% inner_join(life.exp, by = c('id' = 'nuts_id'))
}
\keyword{euro_nuts2}
