% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_county.R
\name{se_county}
\alias{se_county}
\title{Loads a map of Swedish counties for plotting with ggplot2}
\usage{
se_county(keep = 0.05, region = "LAN")
}
\arguments{
\item{keep}{The amount you want to simplify the shapefile to - defaults to five percent}

\item{region}{What you want to call each region when fortified. This should be an area code or other unique identifier.
It defaults to the 2-letter county ID}
}
\description{
This function loads a map of Swedish counties (län), ready for plotting with ggplot. The shapefile comes from Valmyndigheten.se
}
\examples{
sweden <- se_county() \%>\% inner_join(life.exp, by = c('id' = 'county_id'))
}
\keyword{se_county}
