% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_elex_district.R
\name{se_elex_district}
\alias{se_elex_district}
\title{Loads a map of Swedish electoral districts for plotting with ggplot2}
\usage{
se_elex_district(keep = 0.05, region = "VD")
}
\arguments{
\item{keep}{The amount you want to simplify the shapefile to - defaults to five percent}

\item{region}{What you want to call each region when fortified. This should be an area code or other unique identifier.
It defaults to the 8-letter code for each electoral district}
}
\description{
This function loads a map of Swedish electoral districts (valdistrikt), ready for plotting with ggplot. The shapefile comes from Valmyndigheten.se
}
\examples{
sweden <- se_elex_district() \%>\% inner_join(life.exp, by = c('id' = 'district_id'))
}
\keyword{se_county}
