% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jpp.R
\name{theme_jpp}
\alias{theme_jpp}
\title{Add Journalism++ theme to a ggplot2 chart}
\usage{
theme_jpp()
}
\description{
This function allows you to add the Journalism++ theme to your graphics, with theme_jpp()
}
\examples{
line <- ggplot(line_df, aes(x = year, y = lifeExp)) +
geom_line(colour = "#990000", size = 1) +
geom_hline(yintercept = 0, size = 1, colour="#333333") +
theme_jpp()
}
\keyword{theme_jpp}
