% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_map.R
\name{world_map}
\alias{world_map}
\title{Loads a world map}
\usage{
world_map(keep = 0.05, region = "ISO_A3")
}
\arguments{
\item{keep}{The amount you want to simplify the shapefile to - defaults to five percent}

\item{region}{What you want to call each region when fortified. This should be an area code or other unique identifier.
It defaults to the 3-letter ISO code}
}
\description{
This function loads a world map ready for plotting with ggplot.
}
\examples{
world <- world_map() \%>\% inner_join(life.exp, by = c('id' = 'three_letter_iso'))
}
\keyword{world_map}
